refactor: move to foreach rather than for
Minor change - but looks more readable to me (other views and opinions
are valid and wanted!)

As well as this, running count() inside the for block like that results in the
items being counted every time, so if for is to be stuck with, recommend
moving the count outside of that so it's only run once.

Also these sorts of refactors can be more confidently done now there's
tests in place!